feat: 미사용 훅 삭제, 애니메이팅 함수 분리
